move table_or_source to one sub mod
